# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Bora Jogar API
  version: 0.1.0
paths:
...
  /api/v1/admin/dashboard:
    get: { summary: Private beta dashboard, security: [{ cookieAuth: [] }], responses: { '200': { description: Dashboard }, '403': { $ref: '#/components/responses/Forbidden' } } }
...
  /api/v1/admin/users:
    get: { summary: Search users, security: [{ cookieAuth: [] }], parameters: [{ in: query, name: q, schema: { type: string } }], responses: { '200': { description: Users }, '403': { $ref: '#/components/responses/Forbidden' } } }
  /api/v1/admin/matchmaking/runs:
    get: { summary: Inspect matchmaking runs, security: [{ cookieAuth: [] }], responses: { '200': { description: Runs }, '403': { $ref: '#/components/responses/Forbidden' } } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: borajogar_session }
  schemas:
...
    ApiError:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
        requestId: { type: string }
  responses:
...
    Forbidden: { description: Administrator access required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...